refactor: simplify mapping
