+++ apps/loan/src/pages/_app.tsx
@@ -8,8 +8,8 @@ import '@/globals.css'
 import { useCallback, useEffect, useState } from 'react'
 import { HashRouter } from 'react-router-dom'
 import type { AppProps } from 'next/app'
-import { connectWalletLocales } from '@/common/shared/features/connect-wallet'
-import { initOnboard } from '@/common/shared/features/connect-wallet'
+import { connectWalletLocales } from '@/common/features/connect-wallet'
+import { initOnboard } from '@/common/features/connect-wallet'
 import { REFRESH_INTERVAL } from '@/constants'
 import GlobalStyle from '@/globalStyle'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
